# docs/openapi.yml
openapi: 3.1.0
info:
  title: Allama API
  summary: Allama API
  description: Allama is the open source automation platform for enterprise. You can
    operate Allama in headless mode by using the API to create, manage, and run
    workflows.
  termsOfService: https://docs.google.com/document/d/e/2PACX-1vQvDe3SoVAPoQc51MgfGCP71IqFYX_rMVEde8zC4qmBCec5f8PLKQRdxa6tsUABT8gWAR9J-EVs2CrQ/pub
  contact:
    name: Allama Founders
    email: founders@allama.com
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.html
  version: '1'
paths:
...
  /settings/git:
    get:
      tags:
      - settings
      summary: Get Git Settings
      operationId: settings-get_git_settings
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GitSettingsRead'
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
    patch:
      tags:
      - settings
      summary: Update Git Settings
      operationId: settings-update_git_settings
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GitSettingsUpdate'
        required: true
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
  /settings/saml:
    get:
      tags:
      - settings
      summary: Get Saml Settings
      operationId: settings-get_saml_settings
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SAMLSettingsRead'
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
    patch:
      tags:
      - settings
      summary: Update Saml Settings
      operationId: settings-update_saml_settings
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SAMLSettingsUpdate'
        required: true
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
  /settings/auth:
    get:
      tags:
      - settings
      summary: Get Auth Settings
      operationId: settings-get_auth_settings
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSettingsRead'
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
    patch:
      tags:
      - settings
      summary: Update Auth Settings
      operationId: settings-update_auth_settings
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthSettingsUpdate'
        required: true
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
  /settings/oauth:
    get:
      tags:
      - settings
      summary: Get Oauth Settings
      operationId: settings-get_oauth_settings
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthSettingsRead'
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
    patch:
      tags:
      - settings
      summary: Update Oauth Settings
      operationId: settings-update_oauth_settings
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OAuthSettingsUpdate'
        required: true
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
  /settings/app:
    get:
      tags:
      - settings
      summary: Get App Settings
      operationId: settings-get_app_settings
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppSettingsRead'
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
    patch:
      tags:
      - settings
      summary: Update App Settings
      operationId: settings-update_app_settings
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppSettingsUpdate'
        required: true
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
...
components:
  schemas:
...
    AppSettingsRead:
      properties:
        app_registry_validation_enabled:
          type: boolean
          title: App Registry Validation Enabled
        app_executions_query_limit:
          type: integer
          title: App Executions Query Limit
        app_interactions_enabled:
          type: boolean
          title: App Interactions Enabled
      type: object
      required:
      - app_registry_validation_enabled
      - app_executions_query_limit
      - app_interactions_enabled
      title: AppSettingsRead
      description: Settings for the app.
    AppSettingsUpdate:
      properties:
        app_registry_validation_enabled:
          type: boolean
          title: App Registry Validation Enabled
          description: Whether registry validation is enabled.
          default: false
        app_executions_query_limit:
          type: integer
          title: App Executions Query Limit
          description: The maximum number of executions to return in a single query.
          default: 100
        app_interactions_enabled:
          type: boolean
          title: App Interactions Enabled
          description: Whether app interactions are enabled.
          default: false
      type: object
      title: AppSettingsUpdate
      description: Settings for OAuth authentication.
...
    AuthSettingsRead:
      properties:
        auth_basic_enabled:
          type: boolean
          title: Auth Basic Enabled
        auth_require_email_verification:
          type: boolean
          title: Auth Require Email Verification
        auth_allowed_email_domains:
          items:
            type: string
          type: array
          title: Auth Allowed Email Domains
        auth_min_password_length:
          type: integer
          title: Auth Min Password Length
        auth_session_expire_time_seconds:
          type: integer
          title: Auth Session Expire Time Seconds
      type: object
      required:
      - auth_basic_enabled
      - auth_require_email_verification
      - auth_allowed_email_domains
      - auth_min_password_length
      - auth_session_expire_time_seconds
      title: AuthSettingsRead
    AuthSettingsUpdate:
      properties:
        auth_basic_enabled:
          type: boolean
          title: Auth Basic Enabled
          description: Whether basic auth is enabled.
          default: true
        auth_require_email_verification:
          type: boolean
          title: Auth Require Email Verification
          description: Whether email verification is required for authentication.
          default: false
        auth_allowed_email_domains:
          items:
            type: string
          type: array
          title: Auth Allowed Email Domains
          description: Allowed email domains for authentication. If empty, all domains
            are allowed.
        auth_min_password_length:
          type: integer
          title: Auth Min Password Length
          description: Minimum password length for authentication.
          default: 12
        auth_session_expire_time_seconds:
          type: integer
          title: Auth Session Expire Time Seconds
          description: Session expiration time in seconds.
          default: 604800
      type: object
      title: AuthSettingsUpdate
...
    GitSettingsRead:
      properties:
        git_allowed_domains:
          items:
            type: string
          type: array
          title: Git Allowed Domains
        git_repo_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Git Repo Url
        git_repo_package_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Git Repo Package Name
      type: object
      required:
      - git_allowed_domains
      title: GitSettingsRead
    GitSettingsUpdate:
      properties:
        git_allowed_domains:
          items:
            type: string
          type: array
          title: Git Allowed Domains
          description: Allowed git domains for authentication.
        git_repo_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Git Repo Url
        git_repo_package_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Git Repo Package Name
      type: object
      title: GitSettingsUpdate
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    OAuthSettingsRead:
      properties:
        oauth_google_enabled:
          type: boolean
          title: Oauth Google Enabled
      type: object
      required:
      - oauth_google_enabled
      title: OAuthSettingsRead
      description: Settings for OAuth authentication.
    OAuthSettingsUpdate:
      properties:
        oauth_google_enabled:
          type: boolean
          title: Oauth Google Enabled
          description: Whether OAuth is enabled.
          default: true
      type: object
      title: OAuthSettingsUpdate
      description: Settings for OAuth authentication.
...
    SAMLSettingsRead:
      properties:
        saml_enabled:
          type: boolean
          title: Saml Enabled
        saml_enforced:
          type: boolean
          title: Saml Enforced
        saml_idp_metadata_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Saml Idp Metadata Url
        saml_sp_acs_url:
          type: string
          title: Saml Sp Acs Url
      type: object
      required:
      - saml_enabled
      - saml_enforced
      - saml_sp_acs_url
      title: SAMLSettingsRead
    SAMLSettingsUpdate:
      properties:
        saml_enabled:
          type: boolean
          title: Saml Enabled
          description: Whether SAML is enabled.
          default: true
        saml_enforced:
          type: boolean
          title: Saml Enforced
          description: Whether SAML is enforced. If true, users can only use SAML
            to authenticate. Requires SAML to be enabled.
          default: false
        saml_idp_metadata_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Saml Idp Metadata Url
      type: object
      title: SAMLSettingsUpdate
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    APIKeyCookie:
      type: apiKey
      in: cookie
      name: fastapiusersauth
    APIKeyHeader:
      type: apiKey
      in: header
      name: x-allama-service-key
...